Move breadcrumb calcs to a separate file
